bugfix:commonalert
1.suggestrule cli 调整
